Update changelog and correct annotations for entity classes.
