+++ src/main/java/com/synopsys/integration/alert/database/entity/configuration/RegisteredDescriptorEntity.java
@@ -21,7 +21,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package com.synopsys.integration.alert.database.entity.descriptor;
+package com.synopsys.integration.alert.database.entity.configuration;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
@@ -30,34 +30,20 @@ import javax.persistence.Table;
 import com.synopsys.integration.alert.database.entity.DatabaseEntity;
 
 @Entity
-@Table(schema = "ALERT", name = "FIELD_VALUES")
-public class FieldValueEntity extends DatabaseEntity {
-    @Column(name = "CONFIG_ID")
-    private Long configId;
-    @Column(name = "FIELD_ID")
-    private Long fieldId;
-    @Column(name = "FIELD_VALUE")
-    private String value;
+@Table(schema = "ALERT", name = "REGISTERED_DESCRIPTORS")
+public class RegisteredDescriptorEntity extends DatabaseEntity {
+    @Column(name = "NAME")
+    private String name;
 
-    public FieldValueEntity() {
+    public RegisteredDescriptorEntity() {
         // JPA requires default constructor definitions
     }
 
-    public FieldValueEntity(final Long configId, final Long fieldId, final String value) {
-        this.configId = configId;
-        this.fieldId = fieldId;
-        this.value = value;
+    public RegisteredDescriptorEntity(final String name) {
+        this.name = name;
     }
 
-    public Long getConfigId() {
-        return configId;
-    }
-
-    public Long getFieldId() {
-        return fieldId;
-    }
-
-    public String getValue() {
-        return value;
+    public String getName() {
+        return name;
     }
 }
