+++ src/main/java/com/synopsys/integration/alert/database/repository/configuration/DescriptorFieldRepository.java
@@ -5,7 +5,7 @@ import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Component;
 
-import com.synopsys.integration.alert.database.entity.descriptor.DescriptorFieldRelation;
+import com.synopsys.integration.alert.database.entity.configuration.DescriptorFieldRelation;
 import com.synopsys.integration.alert.database.relation.key.DescriptorFieldRelationPK;
 
 @Component
