+++ src/main/java/com/synopsys/integration/alert/database/repository/configuration/FieldValueRepository.java
@@ -28,7 +28,7 @@ import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Component;
 
-import com.synopsys.integration.alert.database.entity.descriptor.FieldValueEntity;
+import com.synopsys.integration.alert.database.entity.configuration.FieldValueEntity;
 
 @Component
 public interface FieldValueRepository extends JpaRepository<FieldValueEntity, Long> {
