+++ src/main/java/com/synopsys/integration/alert/database/repository/configuration/RegisteredDescriptorRepository.java
@@ -28,7 +28,7 @@ import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Component;
 
-import com.synopsys.integration.alert.database.entity.descriptor.RegisteredDescriptorEntity;
+import com.synopsys.integration.alert.database.entity.configuration.RegisteredDescriptorEntity;
 
 @Component
 public interface RegisteredDescriptorRepository extends JpaRepository<RegisteredDescriptorEntity, Long> {
